Triangulation works

+++ src/JSONParser/JSONParser.cpp
@@ -4,11 +4,11 @@ JSONParser::JSONParser()
 {
 }
 
-std::vector<Eigen::Vector2f> JSONParser::getVector(std::string filename)
+std::vector<Eigen::Vector2d> JSONParser::getVector(std::string filename)
 {
   json vec;
   std::ifstream jsonFile(filename);
-  std::vector<Eigen::Vector2f> result;
+  std::vector<Eigen::Vector2d> result;
   if (jsonFile.is_open()) {
     unsigned int size = 0;
     std::string line;
@@ -19,7 +19,7 @@ std::vector<Eigen::Vector2f> JSONParser::getVector(std::string filename)
     jsonFile.close();
     size = vec["size"];
     for(unsigned int i = 0; i < size; i++) {
-      Eigen::Vector2f element(vec["vec"][i][0], vec["vec"][i][1]);
+      Eigen::Vector2d element(vec["vec"][i][0], vec["vec"][i][1]);
       result.push_back(element);
     }
